+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ### Features / Enhancements
 
 - Update to Grafana 9.2.2 (#113)
+- Update CI to upload signed artifacts (#116)
+- Allow to send all or updated only values in Payload (#116)
 
 ## 2.6.0 (2022-10-23)